requirements.txt
PyGithub
pyyaml
gitpython